Refactor build scripts once more
