+++ scripts/boostrap/development.ts
@@ -1,4 +1,4 @@
-import Env, {Environment} from "../app/Env";
+import Env, {Environment} from "../app/libraries/Env";
 
 /**
  * Makes the script crash on unhandled rejections instead of silently
@@ -20,4 +20,3 @@ Env.load();
 const environment: Environment = "development";
 Env.define("BABEL_ENV", environment);
 Env.define("NODE_ENV", environment);
-Env.define("UV_THREADPOOL_SIZE", "128");
